feat(frontend): #1961 add table component for test activity list
