--- frontend/src/views/CONSEP/TestingActivities/TestSearch/constants.ts
@@ -1,153 +1,0 @@
-import ROUTES from '../../../../routes/constants';
-import { CrumbType } from '../../../../components/Breadcrumbs/definitions';
-
-export const DATE_FORMAT = 'Y/m/d';
-
-export const testSearchCrumbs: CrumbType[] = [
-  {
-    name: 'CONSEP',
-    path: ROUTES.CONSEP_FAVOURITE_ACTIVITIES
-  },
-  {
-    name: 'Testing activities search',
-    path: ROUTES.CONSEP_FAVOURITE_ACTIVITIES
-  }
-];
-
-export const testCategoryCodes: string[] = [
-  'PUR',
-  'QA',
-  'QAK',
-  'QAP',
-  'QAR',
-  'QAS',
-  'STD',
-  'TRL'
-];
-
-export const testActivityCodes: string[] = [
-  'AB',
-  'AEX',
-  'AWT',
-  'BIO',
-  'BLN',
-  'BLR',
-  'CAL',
-  'CQA',
-  'CSR',
-  'CUT',
-  'CWR',
-  'D',
-  'D1',
-  'DB2',
-  'DB3',
-  'DRY',
-  'DWD',
-  'DWG',
-  'E17',
-  'E32',
-  'E35',
-  'FUS',
-  'G10',
-  'G11',
-  'G12',
-  'G20',
-  'G30',
-  'G31',
-  'G32',
-  'G34',
-  'G40',
-  'G41',
-  'G42',
-  'G44',
-  'G50',
-  'G52',
-  'G54',
-  'G55',
-  'G56',
-  'G57',
-  'G60',
-  'G62',
-  'G63',
-  'G64',
-  'G70',
-  'G71',
-  'G80',
-  'G90',
-  'GH1',
-  'GSA',
-  'GSE',
-  'HRS',
-  'KLL',
-  'KLN',
-  'LIN',
-  'LSP',
-  'MC',
-  'MCC',
-  'MCK',
-  'MCM',
-  'MCQ',
-  'MCR',
-  'MCW',
-  'MMC',
-  'NGR',
-  'PAT',
-  'PHY',
-  'PLT',
-  'PQA',
-  'PR1',
-  'PR4',
-  'PR8',
-  'PRA',
-  'PRC',
-  'PRE',
-  'PRI',
-  'PUR',
-  'QA',
-  'QM',
-  'QR',
-  'QT',
-  'QU',
-  'RKL',
-  'RKN',
-  'RSK',
-  'RTE',
-  'S10',
-  'S11',
-  'S20',
-  'S31',
-  'S32',
-  'S34',
-  'S44',
-  'S52',
-  'S55',
-  'S56',
-  'S57',
-  'S64',
-  'SA',
-  'SAM',
-  'SC1',
-  'SC2',
-  'SD1',
-  'SDB',
-  'SHP',
-  'SIR',
-  'SOE',
-  'SPB',
-  'SPG',
-  'SPI',
-  'SPL',
-  'SPW',
-  'SRN',
-  'SSP',
-  'SSR',
-  'SW1',
-  'TUM',
-  'TZ',
-  'UPG',
-  'W1',
-  'WDS',
-  'WDU',
-  'WEI',
-  'XRY'
-];
