+++ frontend/src/views/CONSEP/TestingActivities/TestSearch/styles.scss
@@ -1,4 +1,6 @@
 @use '@bcgov-nr/nr-theme/design-tokens/variables.scss' as vars;
+@use '@bcgov-nr/nr-theme/design-tokens/light-theme.scss' as light-theme;
+@use "sass:map";
 
 .consep-test-search-content {
   .consep-test-search-breadcrumb {
@@ -38,5 +40,30 @@
 
   input[type="number"] {
     appearance: textfield;
+  }
+}
+
+.test-list-table-container {
+  margin-top: 2rem;
+  background-color: map.get(light-theme.$light-theme, layer-01);
+
+  .test-list-table-toolbar {
+    margin-inline: 0rem;
+  }
+
+  .test-list-table {
+    margin-inline: 0rem;
+  }
+
+  .test-list-table-toolbar-button {
+    padding: 0.813rem 0.75rem;
+    border-radius: 0rem;
+    border: none;
+    border-left: 1px solid map.get(light-theme.$light-theme, border-subtle-01);
+  }
+
+  .test-list-table-toolbar-button-icon {
+    margin-left: 0.5rem;
+    margin-top: 0.15rem;
   }
 }
